# Behavior chain improvements