--- a/src/NServiceBus.AcceptanceTests/ApiExtension/When_extending_the_publish_api.cs
+++ b/src/NServiceBus.AcceptanceTests/ApiExtension/When_extending_the_publish_api.cs
@@ -63,9 +63,9 @@
                 });
             }
 
-            public class PublishExtensionBehavior : Behavior<IOutgoingLogicalMessageContext>
+            public class PublishExtensionBehavior : IBehavior<IOutgoingLogicalMessageContext, IOutgoingLogicalMessageContext>
             {
-                public override Task Invoke(IOutgoingLogicalMessageContext context, Func<Task> next)
+                public Task Invoke(IOutgoingLogicalMessageContext context, Func<IOutgoingLogicalMessageContext, Task> next)
                 {
                     Context data;
 
@@ -78,7 +78,7 @@
                         Assert.Fail("Expected to find the data");
                     }
 
-                    return next();
+                    return next(context);
                 }
 
                 public class Context