--- a/src/NServiceBus.AcceptanceTests/Basic/When_extending_behavior_context.cs
+++ b/src/NServiceBus.AcceptanceTests/Basic/When_extending_behavior_context.cs
@@ -68,12 +68,12 @@
                 }
             }
 
-            class CustomContextExtensionBehavior : Behavior<IIncomingLogicalMessageContext>
+            class CustomContextExtensionBehavior : IBehavior<IIncomingLogicalMessageContext, IIncomingLogicalMessageContext>
             {
-                public override Task Invoke(IIncomingLogicalMessageContext context, Func<Task> next)
+                public Task Invoke(IIncomingLogicalMessageContext context, Func<IIncomingLogicalMessageContext, Task> next)
                 {
                     context.Extensions.Set("CustomExtension", ExtensionValue);
-                    return next();
+                    return next(context);
                 }
             }
         }