--- a/src/NServiceBus.AcceptanceTests/MessageId/When_message_has_no_id_header.cs
+++ b/src/NServiceBus.AcceptanceTests/MessageId/When_message_has_no_id_header.cs
@@ -21,23 +21,23 @@
             Assert.IsFalse(string.IsNullOrWhiteSpace(context.MessageId));
         }
 
-        public class CorruptionBehavior : Behavior<IDispatchContext>
+        class CorruptionBehavior : IBehavior<IDispatchContext, IDispatchContext>
         {
-            public override Task Invoke(IDispatchContext context, Func<Task> next)
+            public Task Invoke(IDispatchContext context, Func<IDispatchContext, Task> next)
             {
                 context.Operations.First().Message.Headers[Headers.MessageId] = null;
 
-                return next();
+                return next(context);
             }
         }
 
-        public class Context : ScenarioContext
+        class Context : ScenarioContext
         {
             public bool MessageReceived { get; set; }
             public string MessageId { get; set; }
         }
 
-        public class Endpoint : EndpointConfigurationBuilder
+        class Endpoint : EndpointConfigurationBuilder
         {
             public Endpoint()
             {