--- a/src/NServiceBus.AcceptanceTests/Routing/When_publishing_an_interface_with_unobtrusive.cs
+++ b/src/NServiceBus.AcceptanceTests/Routing/When_publishing_an_interface_with_unobtrusive.cs
@@ -35,14 +35,14 @@
                 .Run();
         }
 
-        public class Context : ScenarioContext
+        class Context : ScenarioContext
         {
             public bool GotTheEvent { get; set; }
             public bool Subscribed { get; set; }
             public Type EventTypePassedToRouting { get; set; }
         }
 
-        public class Publisher : EndpointConfigurationBuilder
+        class Publisher : EndpointConfigurationBuilder
         {
             public Publisher()
             {
@@ -60,24 +60,24 @@
                 }).ExcludeType<MyEvent>(); // remove that type from assembly scanning to simulate what would happen with true unobtrusive mode
             }
 
-            class EventTypeSpy : Behavior<IOutgoingLogicalMessageContext>
+            class EventTypeSpy : IBehavior<IOutgoingLogicalMessageContext, IOutgoingLogicalMessageContext>
             {
                 public EventTypeSpy(Context testContext)
                 {
                     this.testContext = testContext;
                 }
 
-                public override Task Invoke(IOutgoingLogicalMessageContext context, Func<Task> next)
+                public Task Invoke(IOutgoingLogicalMessageContext context, Func<IOutgoingLogicalMessageContext, Task> next)
                 {
                     testContext.EventTypePassedToRouting = context.Message.MessageType;
-                    return next();
+                    return next(context);
                 }
 
                 Context testContext;
             }
         }
 
-        public class Subscriber : EndpointConfigurationBuilder
+        class Subscriber : EndpointConfigurationBuilder
         {
             public Subscriber()
             {