--- a/src/NServiceBus.Core.Tests/DataBus/When_applying_the_databus_message_mutator_to_null_properties.cs
+++ b/src/NServiceBus.Core.Tests/DataBus/When_applying_the_databus_message_mutator_to_null_properties.cs
@@ -23,7 +23,7 @@ namespace NServiceBus.Core.Tests.DataBus
                 new BinaryFormatter().Serialize(stream, "test");
                 stream.Position = 0;
 
-                await sendBehavior.Invoke(context, () => TaskEx.CompletedTask);
+                await sendBehavior.Invoke(context, ctx => TaskEx.CompletedTask);
             }
         }