Before I say anything let the data speak

## TL;DR

- Up to 10 (1000%) times faster pipeline execution for the good cases
- Up to 5 (500%) times faster exception handling even when exception is thrown at deepest level
- Gen0 Garbage reduced by 16 times (94%)
- Callstack up to 70% smaller and contains no infrastructure overhead, means dramatic size reduction for error messages sent to Service Control

[Spread sheet with the results ](https://docs.google.com/a/nservicebus.com/spreadsheets/d/1f9DW_EfwJ3P1JfiXg5mMHJxrisKGYlxn40iDfG1d-6M/edit?usp=sharing)

## Optimizations

- `BehaviorChain`, `BehaviorInstance`, `BehaviorInvoker `and Recursion replaced with a statically compiled expression tree which contains the pipeline as if you've written it by hand
- `Behavior<TContext> `is no longer used. Internal behaviors directly implement `IBehavior<TIn, TOut> `which saves a display class allocation and a function delegate allocation
- Outstanding: Replace `ForkConnectors`

## Pipeline Execution Speed
                                           Method |      GarbageCollection |   Calls | PipelineDepth |        Median |    StdDev | Scaled |    Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------------------------------- |----------------------- |-------- |-------------- |-------------- |---------- |------- |--------- |------ |------ |------------------- |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **20000** |            **10** |    **35.3349 ms** | **0.0000 ms** |   **1.00** |        **-** |     **-** |     **-** |      **55'927'608.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   20000 |            10 |     3.0325 ms | 0.0000 ms |   0.09 |        - |     - |     - |         211'610.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   20000 |            10 |    32.9031 ms | 0.0000 ms |   1.00 |     9.00 |     - |     - |      56'013'237.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   20000 |            10 |     2.7232 ms | 0.0000 ms |   0.08 |        - |     - |     - |         248'485.33 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **20000** |            **20** |    **51.7567 ms** | **0.0000 ms** |   **1.00** |        **-** |     **-** |     **-** |     **109'838'378.67** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   20000 |            20 |     5.9089 ms | 0.0000 ms |   0.11 |        - |     - |     - |         426'173.33 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   20000 |            20 |    64.0533 ms | 0.0000 ms |   1.00 |    31.00 |     - |     - |     109'950'981.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   20000 |            20 |     5.7608 ms | 0.0000 ms |   0.09 |        - |     - |     - |         428'402.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **20000** |            **40** |   **105.0577 ms** | **0.0000 ms** |   **1.00** |        **-** |     **-** |     **-** |     **217'655'213.33** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   20000 |            40 |    11.2409 ms | 0.0000 ms |   0.11 |        - |     - |     - |         788'120.00 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   20000 |            40 |   111.1463 ms | 0.0000 ms |   1.00 |    83.00 |     - |     - |     217'746'176.00 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   20000 |            40 |    11.3223 ms | 0.0000 ms |   0.10 |        - |     - |     - |         820'922.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **40000** |            **10** |    **53.2299 ms** | **0.0000 ms** |   **1.00** |        **-** |     **-** |     **-** |     **111'673'618.67** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   40000 |            10 |     6.0083 ms | 0.0000 ms |   0.11 |        - |     - |     - |         211'610.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   40000 |            10 |    64.4422 ms | 0.0000 ms |   1.00 |    32.00 |     - |     - |     111'731'304.00 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   40000 |            10 |     5.6725 ms | 0.0000 ms |   0.09 |        - |     - |     - |         248'485.33 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **40000** |            **20** |   **101.7972 ms** | **0.0000 ms** |   **1.00** |        **-** |     **-** |     **-** |     **219'263'821.33** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   40000 |            20 |    11.8182 ms | 0.0000 ms |   0.12 |        - |     - |     - |         427'752.00 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   40000 |            20 |   119.3070 ms | 0.0000 ms |   1.00 |    83.00 |     - |     - |     219'426'842.67 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   40000 |            20 |    11.4693 ms | 0.0000 ms |   0.10 |        - |     - |     - |         428'402.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **40000** |            **40** |   **197.9773 ms** | **0.0000 ms** |   **1.00** |     **1.00** |     **-** |     **-** |     **434'640'122.67** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   40000 |            40 |    22.0225 ms | 0.0000 ms |   0.11 |        - |     - |     - |         821'712.00 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   40000 |            40 |   214.3654 ms | 0.0000 ms |   1.00 |   185.00 |     - |     - |     434'714'648.00 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   40000 |            40 |    21.9738 ms | 0.0000 ms |   0.10 |        - |     - |     - |         820'922.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **80000** |            **10** |   **112.6092 ms** | **0.0000 ms** |   **1.00** |        **-** |     **-** |     **-** |     **223'107'104.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   80000 |            10 |    11.4271 ms | 0.0000 ms |   0.10 |        - |     - |     - |         213'954.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   80000 |            10 |   116.7316 ms | 0.0000 ms |   1.00 |    85.00 |     - |     - |     223'248'936.00 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   80000 |            10 |    10.9642 ms | 0.0000 ms |   0.09 |        - |     - |     - |         248'485.33 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **80000** |            **20** |   **208.8934 ms** | **0.0000 ms** |   **1.00** |     **1.00** |     **-** |     **-** |     **438'306'378.67** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   80000 |            20 |    23.9429 ms | 0.0000 ms |   0.11 |        - |     - |     - |         426'837.33 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   80000 |            20 |   224.9395 ms | 0.0000 ms |   1.00 |   187.00 |     - |     - |     438'446'882.67 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   80000 |            20 |    23.7259 ms | 0.0000 ms |   0.11 |        - |     - |     - |         428'402.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |   **80000** |            **40** |   **397.3863 ms** | **0.0000 ms** |   **1.00** |     **4.00** |     **-** |     **-** |     **868'527'624.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |   80000 |            40 |    44.7023 ms | 0.0000 ms |   0.11 |        - |     - |     - |         783'466.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |   80000 |            40 |   398.1612 ms | 0.0000 ms |   1.00 |   391.00 |     - |     - |     868'617'648.00 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |   80000 |            40 |    44.4098 ms | 0.0000 ms |   0.11 |        - |     - |     - |         820'922.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **160000** |            **10** |   **217.4958 ms** | **0.0000 ms** |   **1.00** |     **1.00** |     **-** |     **-** |     **446'199'192.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  160000 |            10 |    27.9721 ms | 0.0000 ms |   0.13 |        - |     - |     - |         212'944.00 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  160000 |            10 |   243.0637 ms | 0.0000 ms |   1.00 |   191.00 |     - |     - |     446'298'258.67 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  160000 |            10 |    22.9030 ms | 0.0000 ms |   0.09 |        - |     - |     - |         248'485.33 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **160000** |            **20** |   **431.3198 ms** | **0.0000 ms** |   **1.00** |     **4.00** |     **-** |     **-** |     **876'352'696.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  160000 |            20 |    49.2179 ms | 0.0000 ms |   0.11 |        - |     - |     - |         426'242.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  160000 |            20 |   424.4787 ms | 0.0000 ms |   1.00 |   395.00 |     - |     - |     876'488'970.67 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  160000 |            20 |    47.5439 ms | 0.0000 ms |   0.11 |        - |     - |     - |         428'402.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **160000** |            **40** |   **795.3470 ms** | **0.0000 ms** |   **1.00** |     **6.00** |     **-** |     **-** |   **1'736'349'717.33** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  160000 |            40 |    89.5684 ms | 0.0000 ms |   0.11 |        - |     - |     - |         786'786.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  160000 |            40 |   828.4491 ms | 0.0000 ms |   1.00 |   800.00 |  1.00 |     - |   1'736'454'101.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  160000 |            40 |    90.0768 ms | 0.0000 ms |   0.11 |        - |     - |     - |         820'922.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **320000** |            **10** |   **451.5385 ms** | **0.0000 ms** |   **1.00** |     **4.00** |     **-** |     **-** |     **892'298'410.67** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  320000 |            10 |    47.5526 ms | 0.0000 ms |   0.11 |        - |     - |     - |         213'954.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  320000 |            10 |   472.5423 ms | 0.0000 ms |   1.00 |   403.00 |     - |     - |     892'383'181.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  320000 |            10 |    43.7060 ms | 0.0000 ms |   0.09 |        - |     - |     - |         248'485.33 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **320000** |            **20** |   **834.1605 ms** | **0.0000 ms** |   **1.00** |     **7.00** |     **-** |     **-** |   **1'752'423'080.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  320000 |            20 |    91.0310 ms | 0.0000 ms |   0.11 |        - |     - |     - |         424'901.33 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  320000 |            20 |   829.2922 ms | 0.0000 ms |   1.00 |   809.00 |     - |     - |   1'752'538'165.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  320000 |            20 |    93.8465 ms | 0.0000 ms |   0.11 |        - |     - |     - |         428'402.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **320000** |            **40** | **1,591.8079 ms** | **0.0000 ms** |   **1.00** |    **14.00** |     **-** |     **-** |   **3'471'944'328.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  320000 |            40 |   184.9468 ms | 0.0000 ms |   0.12 |        - |     - |     - |         823'010.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  320000 |            40 | 1,699.7604 ms | 0.0000 ms |   1.00 | 1,622.00 |  1.00 |     - |   3'472'127'090.67 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  320000 |            40 |   179.9737 ms | 0.0000 ms |   0.11 |        - |     - |     - |         820'922.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **640000** |            **10** |   **862.8742 ms** | **0.0000 ms** |   **1.00** |     **7.00** |     **-** |     **-** |   **1'784'558'480.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  640000 |            10 |    95.0701 ms | 0.0000 ms |   0.11 |        - |     - |     - |         211'517.33 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  640000 |            10 |   845.6103 ms | 0.0000 ms |   1.00 |   825.00 |     - |     - |   1'784'552'106.67 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  640000 |            10 |    89.8122 ms | 0.0000 ms |   0.11 |        - |     - |     - |         248'485.33 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **640000** |            **20** | **1,701.0697 ms** | **0.0000 ms** |   **1.00** |    **13.00** |     **-** |     **-** |   **3'504'481'168.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  640000 |            20 |   198.5414 ms | 0.0000 ms |   0.12 |        - |     - |     - |         427'506.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  640000 |            20 | 1,638.1399 ms | 0.0000 ms |   1.00 | 1,639.00 |  1.00 |     - |   3'504'670'221.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  640000 |            20 |   199.7631 ms | 0.0000 ms |   0.12 |        - |     - |     - |         428'402.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |  **640000** |            **40** | **3,751.5274 ms** | **0.0000 ms** |   **1.00** |    **29.00** |     **-** |     **-** |   **6'943'054'322.67** |
  V6_PipelineAfterOptimizations |      Concurrent Server |  640000 |            40 |   376.0074 ms | 0.0000 ms |   0.10 |        - |     - |     - |         788'120.00 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |  640000 |            40 | 3,659.6397 ms | 0.0000 ms |   1.00 | 3,265.00 |  2.00 |     - |   6'943'443'709.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |  640000 |            40 |   378.0525 ms | 0.0000 ms |   0.10 |        - |     - |     - |         820'922.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** | **1280000** |            **10** | **2,740.1182 ms** | **0.0000 ms** |   **1.00** |    **15.00** |     **-** |     **-** |   **3'568'757'013.33** |
  V6_PipelineAfterOptimizations |      Concurrent Server | 1280000 |            10 |   181.1852 ms | 0.0000 ms |   0.07 |        - |     - |     - |         211'424.00 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation | 1280000 |            10 | 2,149.5577 ms | 0.0000 ms |   1.00 | 1,672.00 |     - |     - |   3'568'863'272.00 |
  V6_PipelineAfterOptimizations | Concurrent Workstation | 1280000 |            10 |   179.1115 ms | 0.0000 ms |   0.08 |        - |     - |     - |         248'485.33 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** | **1280000** |            **20** | **4,174.6157 ms** | **0.0000 ms** |   **1.00** |    **29.00** |     **-** |     **-** |   **7'008'672'080.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server | 1280000 |            20 |   372.1575 ms | 0.0000 ms |   0.09 |        - |     - |     - |         427'725.33 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation | 1280000 |            20 | 4,185.4200 ms | 0.0000 ms |   1.00 | 3,298.00 |  2.00 |     - |   7'008'864'472.00 |
  V6_PipelineAfterOptimizations | Concurrent Workstation | 1280000 |            20 |   374.6501 ms | 0.0000 ms |   0.09 |        - |     - |     - |         428'402.67 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** | **1280000** |            **40** | **8,093.3751 ms** | **0.0000 ms** |   **1.00** |    **60.00** |     **-** |     **-** |  **13'885'698'672.00** |
  V6_PipelineAfterOptimizations |      Concurrent Server | 1280000 |            40 |   709.0813 ms | 0.0000 ms |   0.09 |        - |     - |     - |         785'938.67 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation | 1280000 |            40 | 8,130.4504 ms | 0.0000 ms |   1.00 | 6,548.00 |  7.00 |     - |  13'886'089'957.33 |
  V6_PipelineAfterOptimizations | Concurrent Workstation | 1280000 |            40 |   724.6494 ms | 0.0000 ms |   0.09 |        - |     - |     - |         820'922.67 |

[Benchmark used](https://github.com/danielmarbach/MicroBenchmarks/blob/master/MicroBenchmarks/NServiceBus/PipelineExecution.cs)

## Exception handling speed

                         Method |      GarbageCollection | PipelineDepth |     Median |     StdDev | Scaled | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
------------------------------- |----------------------- |-------------- |----------- |----------- |------- |------ |------ |------ |------------------- |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |            **10** | **31.4081 us** |  **3.0626 us** |   **1.00** |     **-** |     **-** |     **-** |           **1'866.52** |
  V6_PipelineAfterOptimizations |      Concurrent Server |            10 | 11.2087 us |  1.4214 us |   0.36 |     - |     - |     - |             358.39 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |            10 | 27.4915 us |  3.6642 us |   1.00 | 11.51 |     - |     - |           2'135.67 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |            10 | 10.7383 us |  1.9679 us |   0.39 |     - |     - |     - |             348.59 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |            **20** | **48.5027 us** |  **7.6032 us** |   **1.00** |     **-** |     **-** |     **-** |           **3'986.92** |
  V6_PipelineAfterOptimizations |      Concurrent Server |            20 | 11.9111 us |  2.4094 us |   0.25 |     - |     - |     - |             347.79 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |            20 | 51.8343 us | 13.0591 us |   1.00 |     - |     - |     - |           3'930.75 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |            20 | 11.5575 us |  1.1563 us |   0.22 |     - |     - |     - |             344.80 |
 **V6_PipelineBeforeOptimizations** |      **Concurrent Server** |            **40** | **80.4565 us** | **12.2315 us** |   **1.00** |     **-** |     **-** |     **-** |           **5'684.50** |
  V6_PipelineAfterOptimizations |      Concurrent Server |            40 | 13.1159 us |  0.6705 us |   0.16 |     - |     - |     - |             325.70 |
 V6_PipelineBeforeOptimizations | Concurrent Workstation |            40 | 83.5192 us | 16.6089 us |   1.00 |     - |     - |     - |           6'054.96 |
  V6_PipelineAfterOptimizations | Concurrent Workstation |            40 | 10.9396 us |  1.6605 us |   0.13 |     - |     - |     - |             351.89 |

[Benchmark used](https://github.com/danielmarbach/MicroBenchmarks/blob/master/MicroBenchmarks/NServiceBus/PipelineException.cs)

## Callstack (based on example of a Publish)

### Before

Approx 113 lines

![callstackbefore](https://cloud.githubusercontent.com/assets/174258/18367017/69cd5b14-7619-11e6-82ca-be65e06f7356.png)

### After

Approx 30 lines, means roughly 73% call stack improvement

![callstackafter](https://cloud.githubusercontent.com/assets/174258/18367036/80b99a9a-7619-11e6-8c55-9781f5240047.png)

## Potential downside

Increases startup time slightly

                         Method |      GarbageCollection | PipelineDepth |         Median |        StdDev | Scaled | Gen 0 | Gen 1 | Gen 2 | Bytes Allocated/Op |
…