@bording 2016-09-08
I think the numbers speak for themselves here, this is a great improvement, with apparently no downsides!