@indualagarsamy 2016-09-08
@danielmarbach - there was an issue a long time ago regarding the stack trace and the exception info. Can you check that issue and make sure we won't negatively impact? I'll try and find that issue. And I think @SimonCropp addressed it.