@indualagarsamy 2016-09-08
@danielmarbach - See related issues:
https://github.com/Particular/NServiceBus/issues/1748
https://github.com/Particular/NServiceBus/issues/2248