@bording 2016-09-09
@danielmarbach I'm assuming there's nothing special required here regarding if the behavior actually has an await vs just returning a task? In other words, what you've got here already handles that correctly?