@andreasohlund 2016-09-09
@indualagarsamy that issue was about us "corrupting"/lying about the callstack when logging. This  PRis ok since here we've made the real callstack smaller