@andreasohlund 2016-09-09
R.I.P callstack cleaning sample