@danielmarbach 2016-09-09
Agree with @simoncropp. The async call stack is still there. So the sample
is still very valid