@mikeminutillo 2016-09-09
So, no else seems to want to be "that guy" so I'll do it. 

Should this wait until after the V6 release? It's really cool but highly invasive and potentially risky. We have a stable release that we're preparing to put out. We're happy with the performance of it (we've done extensive testing that says so). 

Should this be 6.0.1?