@danielmarbach 2016-09-09
@mikeminutillo that is a very valid comment. I had this conversation with @Particular/nservicebus-maintainers I'll add the more context to the description and ping you afterwards