@mikeminutillo 2016-09-09
Also, I suspect this breaks my hacky plugin that diagrams the pipeline but that's not really relevant :)