@danielmarbach 2016-09-09
@mikeminutillo show me the code and I think I can give you a way to retrieve it or we can make the code support it if needed.

Btw. updated the description, see Should this go in v6 or v6.x