@danielmarbach 2016-09-09
@bording about your question https://github.com/Particular/NServiceBus/pull/4125#issuecomment-245807616 yes that should work. From a calling perspective for the generated code a behavior with an invoke method marked as async is not different. The state machine is only relevant for the code in the Invoke method.