@timbussmann 2016-09-09
very impressive!

![cyn2wakueaawgtg](https://cloud.githubusercontent.com/assets/3524870/18386046/847ebf86-7694-11e6-80ba-adf3eb35925a.jpg)

> Should this wait until after the V6 release? It's really cool but highly invasive and potentially risky.

I'm not sure about that part, there is no breaking change and we have hundreds of tests running the pipeline. Releasing this as a 6.0.1, doesn't reduce any risks?