@adamralph 2016-09-10
IMHO the risk of putting this into 6.0 is negligible. If we choose to do that, and we want to remain true the meaning of "release candidate" then it would need to go into RC1. I do not believe we should release RC1 until everything that is planned for 6.0 is in it, otherwise it is *not* a release candidate.

Minor point (no pun intended): if we decide to hold off until post 6.0, then according to SemVer this would induce a minor version bump (it's not a bug fix).