@andreasohlund 2016-09-12
>"release candidate" then it would need to go into RC1

That ship has sailed since RC1 is already on myget (caching hell). So we would just say RC1 was an internal version, RC2 will be the public "RC". This would not cause any churn in the downstreams since this is a non breaking change and they all have nuspec range: `> 6.0.0-rc0001` ?