@adamralph 2016-09-12
Agreed. :+1: to go straight to RC2.