@danielmarbach 2016-09-12
@Particular/nservicebus-maintainers removed WIP, build green. Please review and merge