@timbussmann 2016-09-12
LGTM as well :+1: Let's have @bording have a look too before merging