@bording 2016-09-12
@danielmarbach With the changes introduced in c3d26577d49101699e6fd8097543744a0507268f, that means this change does actually have some public-facing changes, since those abstract classes are public.

Is that something everyone is ok with?

Also, by factoring out the Fork method, you reintroduced a delegate allocation to those abstract classes, after having just gotten rid of it.

I'm assuming the goal here is to get rid of the abstract classes (and the unused extension methods that reference them) altogether, but we aren't doing it now because of public breaks? If the intention is to make sure they aren't used, then it's probably ok for them to keep using the Fork method, even if it does allocate.