@danielmarbach 2016-09-13
@bording all the changes I did here are binary compatible. ApiApprover has a bug and changes the signature because it also incorporates internal interfaces into the public API definition. My plan is to send in a PR against ApiApprover, but this is going to be treated outside this PR.

![image](https://cloud.githubusercontent.com/assets/174258/18465472/4c4f615a-7997-11e6-9f4b-c0ba15a58bde.png)