@andreasohlund 2016-09-13
@bording 's comments seems to have been addressed, will merge once green.